New drawer folder made for cleaner file structure within the Navbar folder
